플레이어 접속, 이동 구현 (나중에 들어온 플레이어 동기화는 아직)
나중에 들어온 플레이어에게도 기존에 있던 플레이어의 정보를 주어주기위해서 서버에서 플레이어들의 정보를 저장할 필요성을 느낌

위 사항 수정 필요, 코드 정리 및 주석 필요(너무 구현에만 신경쓰느라 주석을 달지않음 + 코드가 난잡함)

+++ TCPServerWithGUI/TCPServerWithGUI/TCPServerCore.h
@@ -13,7 +13,7 @@ public:
 
 	std::unordered_map<unsigned short, Client>& GetClients();
 
-	short AddClient(SOCKET clientSocket);
+	byte AddClient(SOCKET clientSocket);
 	void ExitClient(unsigned short id);
 
 private:
